🧑‍💻 🍻 🚧 (UI): Cleaned-up workspace to improve productivity
