+++ packages/client/lib/PrototypeFiles/main.dart
@@ -1,0 +1,44 @@
+/* import 'package:flutter/material.dart';
+import 'package:client/login_page.dart';
+import 'package:client/home.dart';
+import 'package:client/searchhistory.dart';
+import 'package:client/viewAllProblems.dart';
+
+List<String> history = [];
+
+// void main() {
+//   runApp(MaterialApp(
+//     home: Home(history: history),
+//     routes: {
+//       '/home.dart': (context) => Home(history: history),
+//       '/searchhistory.dart': (context) => SearchHistory(history),
+//       '/viewAllProblems.dart': (context) => viewAll(),
+//     },
+//   ));
+// }
+
+void main() {
+  runApp(const MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+        title: 'MathU Similarity Index',
+        debugShowCheckedModeBanner: false,
+        theme: ThemeData(
+          primarySwatch: Colors.blue,
+        ),
+        home: const SignUpPage(),
+        routes: {
+          '/home.dart': (context) => Home(history: history),
+          '/searchhistory.dart': (context) => SearchHistory(history),
+          '/viewAllProblems.dart': (context) => viewAll(),
+        });
+  }
+}
+ */
