+++ packages/client/lib/PrototypeFiles/search_bar.dart
