+++ packages/client/lib/PrototypeFiles/searchhistory.dart
@@ -1,4 +1,4 @@
-//REFERENCING IS AT BOTTOM OF PAGE
+/* //REFERENCING IS AT BOTTOM OF PAGE
 import 'package:flutter/material.dart';
 import 'package:client/home.dart';
 import 'package:client/viewAllProblems.dart';
@@ -130,3 +130,4 @@ CONSTRUCTORS DART/FLUTTER TUTORIALS -> Dart/Flutter Constructors tutorial with e
                                         Calling a function with an array of parameters in Dart URL(https://stackoverflow.com/questions/42965731/calling-a-function-with-an-array-of-parameters-in-dart)
                                         ARRAYS IN FLUTTER URL(https://www.fluttercampus.com/guide/198/array-list-basics-dart-flutter/)
 */
+ */
